fix: restore terminal profiles across app restart
